Refine TUI settings flow
